fix: jpa build

+++ store-manager-api/src/main/kotlin/com/rtr/store_manager_api/dto/InventoryMovementRequestDTO.kt
@@ -1,11 +1,13 @@
 package com.rtr.store_manager_api.dto
 
+import java.math.BigDecimal
+
 data class InventoryMovementRequestDTO(
     val productId: String,
     val userId: String,
     val quantity: Int,
-    val unitPurchasePrice: Double? = null,
-    val unitSalePrice: Double? = null,
+    val unitPurchasePrice: BigDecimal? = null,
+    val unitSalePrice: BigDecimal? = null,
     val type: String,
     val description: String? = null
 )
